requirements.txt
attrs==23.1.0
Automat==22.10.0
awscli==1.29.22
beautifulsoup4==4.12.2
boto3==1.28.22
botocore==1.31.22
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.4
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
docutils==0.16
filelock==3.12.2
ghp-import==2.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
mergedeep==1.3.4
mkdocs==1.5.2
mkdocs-material==9.1.21
mkdocs-material-extensions==1.1.1
packaging==23.1
parsel==1.8.1
pathspec==0.11.2
platformdirs==3.10.0
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.16.1
pymdown-extensions==10.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
PyYAML==6.0.1
pyyaml-env-tag==0.1
queuelib==1.6.2
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.6.1
Scrapy==2.10.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
tldextract==3.4.4
Twisted==22.10.0
typing-extensions==4.7.1
urllib3==1.26.16
w3lib==2.1.2
watchdog==3.0.0
zipp==3.16.2
zope.interface==6.0
